Implemented some features
- Added Shoot from player and autoAim
- Added correct handling of ScriptableObjects
- Added projectile damage on enemy and player
- Added "Load next room" on all enemies death
- Added projectile shoot handling

+++ Assets/Scripts/Gameplay/Definitions/Entity.cs
@@ -16,7 +16,7 @@ public class Entity : ScriptableObject
 [Serializable]
 public class EntityStats
 {
-    public int m_HP;
+    public float m_HP;
     public int m_Damage;
     public int m_Speed;
     public int m_ShotSpeed;
